Added exception handling

+++ CountriesApp/CountriesApp.API/Controllers/CountriesController.cs
@@ -7,6 +7,7 @@ using System.Threading.Tasks;
 using CountriesApp.BLL;
 using CountriesApp.BLL.Interfaces;
 using CountriesApp.BLL.Models;
+using CountriesApp.API.Tools;
 
 namespace CountriesApp.API.Controllers
 {
@@ -15,10 +16,12 @@ namespace CountriesApp.API.Controllers
     public class CountriesController : ControllerBase
     {
         private ICountriesBLL _BLL;
+        private CountriesControllerHelper _countriesControllerHelper;
 
         public CountriesController()
         {
             _BLL = new CountriesBLL();
+            _countriesControllerHelper = new CountriesControllerHelper();
         }
 
         /// <summary>
@@ -35,9 +38,10 @@ namespace CountriesApp.API.Controllers
         /// <param name="country"></param>
         [HttpPost]
         [Route("addCountry")]
-        public void AddCountry([FromBody] CountryDTO country)
+        public IActionResult AddCountry([FromBody] CountryDTO country)
         {
             _BLL.AddCountry(country);
+            return Ok();
         }
 
         /// <summary>
@@ -47,7 +51,17 @@ namespace CountriesApp.API.Controllers
         /// <returns>Existing Country instance</returns>
         [HttpGet]
         [Route("getCountryById")]
-        public ActionResult<CountryDTO> GetCountry(Guid id) => Ok(_BLL.GetCountry(id));
+        public ActionResult<CountryDTO> GetCountry(Guid id)
+        {
+            var country = _BLL.GetCountry(id);
+
+            if (!_countriesControllerHelper.IsCountry(country))
+            {
+                return NotFound("There is no country in the database with such id.");
+            }
+
+            return Ok(country);
+        }
 
         /// <summary>
         /// Finds country in database by its name.
@@ -56,7 +70,17 @@ namespace CountriesApp.API.Controllers
         /// <returns>Existing Country instance</returns>
         [HttpGet]
         [Route("getCountryByName")]
-        public ActionResult<CountryDTO> GetCountry(string name) => Ok(_BLL.GetCountry(name));
+        public ActionResult<CountryDTO> GetCountry(string name)
+        {
+            var country = _BLL.GetCountry(name);
+
+            if (!_countriesControllerHelper.IsCountry(country))
+            {
+                return NotFound("There is no country in the database with such name.");
+            }
+
+            return Ok(country);
+        }
 
         /// <summary>
         /// Updates existing Country instance.
@@ -64,9 +88,10 @@ namespace CountriesApp.API.Controllers
         /// <param name="country"></param>
         [HttpPut]
         [Route("updateCountry")]
-        public void UpdateCountry(CountryDTO country)
+        public IActionResult UpdateCountry(CountryDTO country)
         {
             _BLL.UpdateCountry(country);
+            return Ok();
         }
 
         /// <summary>
@@ -75,9 +100,17 @@ namespace CountriesApp.API.Controllers
         /// <param name="id"></param>
         [HttpDelete]
         [Route("deleteCountry")]
-        public void DeleteCountry(Guid id)
+        public IActionResult DeleteCountry(Guid id)
         {
+            var countryToDelete = _BLL.GetCountry(id);
+
+            if (!_countriesControllerHelper.IsCountry(countryToDelete))
+            {
+                return NotFound("There is no country in the database with such id.");
+            }
+
             _BLL.DeleteCountry(id);
+            return Ok();
         }
     }
 }
